create_dataset/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
comet_ml==3.55.0
configobj==5.0.9
dulwich==0.24.10
entmax==1.3
everett==3.1.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-translate==3.23.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonargparse==3.13.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lightning-utilities==0.15.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.3
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.47.0
simplejson==3.20.2
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
torch==2.9.1
torchmetrics==0.10.3
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
unbabel-comet @ git+https://github.com/Unbabel/COMET.git@da61b810ea06f67bbd5d96031f55082167a16e80
urllib3==2.6.2
wrapt==2.0.1
wurlitzer==3.1.1
yarl==1.22.0
